fix merge sorta

+++ src/main/java/com/kytheralabs/SeleniumScripter.java
@@ -301,8 +301,8 @@ public class SeleniumScripter {
     private void runWait(Map<String, Object> script) throws Exception {
         int waittimeout = 180;
         if(script.containsKey("timeout")){
-            //waittimeout = ((Double) script.get("timeout")).intValue();
-            waittimeout = (Integer) script.get("timeout");
+            waittimeout = ((Double) script.get("timeout")).intValue();
+            //waittimeout = (Integer) script.get("timeout");
         }
 
         if(script.get("selector").toString().equals("none")){
